# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aigoni API
  version: 1.0.0
  description: |
    Aigoni API contract. `x-implementation-status` marks routes registered now
    implemented REST contract.
servers:
  - url: /
...
paths:
...
  /rest/v1/categories:
    get:
      tags: [Public REST]
      summary: List published post categories
      description: Includes a built-in uncategorized item whenever published posts exist.
      x-implementation-status: implemented
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
        '500':
          $ref: '#/components/responses/Error'
  /rest/v1/posts:
    get:
      tags: [Public REST]
      summary: List published posts
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: category
          in: query
          schema: {type: string}
          description: Exact category filter. Use __none__ for posts with an empty category. Implemented.
        - name: tag
          in: query
          schema: {type: string}
          description: Exact tag filter. Implemented.
        - name: year
          in: query
          schema: {type: string, pattern: '^[0-9]{4}$'}
          description: Four-digit year filter. Implemented.
      responses:
        '200':
          description: Post page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicContentList'
        '500':
          $ref: '#/components/responses/Error'
  /rest/v1/posts/{slug}:
    get:
      tags: [Public REST]
      summary: Get a published post
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Post detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicContent'
        '404':
          $ref: '#/components/responses/Error'
  /rest/v1/pages:
    get:
      tags: [Public REST]
      summary: List published pages
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Page list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicContentList'
  /rest/v1/pages/{slug}:
    get:
      tags: [Public REST]
      summary: Get a published page
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Page detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicContent'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: aigoni_session
      description: HttpOnly, SameSite=Lax; Secure over HTTPS.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on Admin write operations in the target contract.
    APIKeyHeader: {type: apiKey, in: header, name: X-API-Key}
    BearerAuth: {type: http, scheme: bearer}
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PerPage: {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 10}}
...
    Slug: {name: slug, in: path, required: true, schema: {type: string}}
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
...
    PublicContent:
      type: object
      required: [title, description, date, lastmod, slug, tags, url]
      properties:
        title: {type: string}
        description: {type: string}
        date: {type: string, format: date-time}
        lastmod: {type: string, format: date-time}
        slug: {type: string}
        category: {type: string}
        tags: {type: array, items: {type: string}}
        cover_image: {type: string}
        toc: {type: boolean}
        body: {type: string, description: Present in detail responses.}
        html: {type: string, description: Present in detail responses.}
        excerpt: {type: string, description: Present in search results.}
        url: {type: string}
    PublicContentList: {type: object, required: [items, page, per_page, total, total_pages], properties: {items: {type: array, items: {$ref: '#/components/schemas/PublicContent'}}, page: {type: integer}, per_page: {type: integer}, total: {type: integer}, total_pages: {type: integer}}}
    Category: {type: object, required: [name, count, url], properties: {name: {type: string}, count: {type: integer}, url: {type: string}, none: {type: boolean, description: Present only on the built-in uncategorized item.}}}
    CategoryList: {type: object, required: [items, total], properties: {items: {type: array, items: {$ref: '#/components/schemas/Category'}}, total: {type: integer}}}
...
    ErrorResponse: {type: object, required: [error], properties: {error: {oneOf: [{type: string}, {$ref: '#/components/schemas/ErrorDetail'}]}}}
    ErrorDetail: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}, field_errors: {type: object, additionalProperties: {type: string}}, request_id: {type: string}}}